final fixes in onboarding layout
