+++ frontend/src/constants/steps.ts
@@ -2,4 +2,5 @@ export const STEPS = {
   AVATAR_SELECTION_STEP: 'avatarSelection',
   FOLDER_SETUP_STEP: 'folderSetup',
   THEME_SELECTION_STEP: 'themeSelection',
+  UPDATE_STEP: 'updates',
 };
